refactor: rename alerts context and update data handling in AlertsListViewMap for improved clarity and functionality
