+++ apps/frontend/src/components/alerts/AlertsListViewMap/index.tsx
@@ -5,7 +5,7 @@
 import { Surface } from '@/components/layout/Surface';
 import { MapView } from '@/components/map/MapView';
 import { MapViewStyleAlerts, MapViewStyleAlertsLayerId } from '@/components/map/MapViewStyleAlerts';
-import { useAlertsListContext } from '@/contexts/AlertsList.context';
+import { useAlertsContext } from '@/contexts/Alerts.context';
 import { useEnvironmentContext } from '@/contexts/Environment.context';
 import { centerMap } from '@/utils/map.utils';
 import * as turf from '@turf/turf';
@@ -23,18 +23,17 @@ export function AlertsListViewMap() {
 
 	const { alertsListMap } = useMap();
 	const router = useRouter();
-	const alertsListContext = useAlertsListContext();
+	const alertsContext = useAlertsContext();
 	const environmentContext = useEnvironmentContext();
-
 	//
 	// B. Handle actions
 
 	useEffect(() => {
 		// Exit early if there are no alerts or map
-		if (!alertsListContext.data.filtered_fc || !alertsListContext.data.filtered_fc.features.length || !alertsListMap) return;
+		if (!alertsContext.data.featureCollection || !alertsContext.data.featureCollection.features.length || !alertsListMap) return;
 
 		// When there are search filters, center the map on the cluster with the most points
-		const clusterPoints = turf.clustersKmeans(alertsListContext.data.filtered_fc, { mutate: true, numberOfClusters: 2 });
+		const clusterPoints = turf.clustersKmeans(alertsContext.data.featureCollection, { mutate: true, numberOfClusters: 2 });
 		const clusterPointsCount = clusterPoints.features.reduce((acc, feature) => {
 			if (typeof feature.properties.cluster !== 'number') return acc;
 			const clusterId = feature.properties.cluster;
@@ -46,7 +45,7 @@ export function AlertsListViewMap() {
 		const filteredClusterPoints = clusterPoints.features.filter(feature => feature.properties.cluster === Number(clusterId));
 		centerMap(alertsListMap, filteredClusterPoints);
 		//
-	}, [alertsListContext.data.filtered_fc, alertsListMap]);
+	}, [alertsContext.data.featureCollection, alertsListMap]);
 
 	function handleLayerClick(event: MapMouseEvent) {
 		const feature = event.features?.[0];
@@ -66,7 +65,7 @@ export function AlertsListViewMap() {
 					interactiveLayerIds={[MapViewStyleAlertsLayerId]}
 					onClick={handleLayerClick}
 				>
-					<MapViewStyleAlerts data={alertsListContext.data.filtered_fc} />
+					<MapViewStyleAlerts data={alertsContext.data.featureCollection} />
 				</MapView>
 			</div>
 		</Surface>
